Refactor modules for the notifications screen
