+++ frontend/components/NotificationComponent/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import {StyleSheet, View} from 'react-native';
+import {View} from 'react-native';
 import {Avatar, Button, Divider, Text} from 'react-native-paper';
+import styles from './styles';
 
 /*
  * Props:
@@ -40,34 +41,5 @@ function NotificationComponent(props) {
     </View>
   );
 }
-
-// Styles
-const styles = StyleSheet.create({
-  view: {
-    flexDirection: 'row',
-    marginTop: 12
-  },
-
-  notification_avatar: {
-    marginHorizontal: 12
-  },
-
-  button: {
-    width: 100,
-    marginEnd: 40,
-    marginVertical: 12
-  },
-
-  button_positive: {
-  },
-
-  button_negative: {
-    backgroundColor: 'grey'
-  },
-
-  buttons_view: {
-    flexDirection: 'row'
-  }
-});
 
 export default NotificationComponent;
