+++ frontend/components/NotificationComponent/styles.js
@@ -1,0 +1,31 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+  view: {
+    flexDirection: 'row',
+    marginTop: 12
+  },
+
+  notification_avatar: {
+    marginHorizontal: 12
+  },
+
+  button: {
+    width: 100,
+    marginEnd: 40,
+    marginVertical: 12
+  },
+
+  button_positive: {
+  },
+
+  button_negative: {
+    backgroundColor: 'grey'
+  },
+
+  buttons_view: {
+    flexDirection: 'row'
+  }
+});
+
+export default styles;
